Decomposed HomeworkItem in a separate component for students as preparation for teacher items
