+++ src/components/molecules/vHomeworkItemStudent.vue
@@ -1,0 +1,96 @@
+<template>
+	<v-list-item :key="homework._id" :href="homeworkHref(homework._id)">
+		<v-list-item-avatar class="hidden-xs-only">
+			<img :src="taskIconSvg" role="presentation" />
+		</v-list-item-avatar>
+		<v-list-item-content>
+			<v-list-item-subtitle class="text-wrap">
+				{{ homework.courseName }}
+			</v-list-item-subtitle>
+			<v-list-item-title
+				class="text-wrap"
+				v-text="homework.name"
+			></v-list-item-title>
+		</v-list-item-content>
+		<v-list-item-action>
+			<v-list-item-action-text
+				class="subtitle-2"
+				data-test-id="dueDateLabel"
+				v-text="
+					computedDueDateLabel(
+						homework.duedate,
+						(shorten = $vuetify.breakpoint.xsOnly)
+					)
+				"
+			/>
+			<v-spacer />
+			<v-custom-chip-time-remaining
+				v-if="isCloseToDueDate(homework.duedate)"
+				type="warning"
+				:due-date="homework.duedate"
+				:shorten-date="$vuetify.breakpoint.xsOnly"
+			/>
+			<v-custom-chip-time-remaining
+				v-else-if="isOverDue(homework.duedate)"
+				type="overdue"
+				:due-date="homework.duedate"
+				:shorten-date="$vuetify.breakpoint.xsOnly"
+			/>
+		</v-list-item-action>
+	</v-list-item>
+</template>
+
+<script>
+import VCustomChipTimeRemaining from "@components/molecules/VCustomChipTimeRemaining";
+import taskIconSvg from "@assets/img/courses/task-new.svg";
+import { fromNow, fromNowToFuture } from "@plugins/datetime";
+import {
+	printDateFromStringUTC,
+	printDateTimeFromStringUTC,
+} from "@plugins/datetime";
+
+export default {
+	components: { VCustomChipTimeRemaining },
+	props: {
+		homework: {
+			type: Object,
+			required: true,
+		},
+	},
+	data() {
+		return {
+			fromNow,
+			taskIconSvg,
+		};
+	},
+	methods: {
+		computedDueDateLabel(dueDate, shorten = false) {
+			if (!dueDate) {
+				return this.$t("pages.homeworks.labels.noDueDate");
+			} else if (shorten) {
+				return (
+					this.$t("pages.homeworks.labels.due") +
+					printDateFromStringUTC(dueDate)
+				);
+			} else {
+				return (
+					this.$t("pages.homeworks.labels.due") +
+					printDateTimeFromStringUTC(dueDate)
+				);
+			}
+		},
+		isCloseToDueDate(dueDate) {
+			const timeDiff = fromNowToFuture(dueDate, "hours");
+			if (timeDiff === null) {
+				return false;
+			} else return timeDiff <= 24;
+		},
+		isOverDue(dueDate) {
+			return dueDate && new Date(dueDate) < new Date();
+		},
+		homeworkHref: (id) => {
+			return "/homework/" + id;
+		},
+	},
+};
+</script>
