+++ src/components/organisms/HomeworksList.vue
@@ -14,62 +14,25 @@
 		</template>
 
 		<template v-for="(homework, index) of homeworks" v-else>
-			<v-list-item :key="homework._id" :href="homeworkHref(homework._id)">
-				<v-list-item-avatar class="hidden-xs-only">
-					<img :src="taskIconSvg" role="presentation" />
-				</v-list-item-avatar>
-				<v-list-item-content>
-					<v-list-item-subtitle class="text-wrap">
-						{{ homework.courseName }}
-					</v-list-item-subtitle>
-					<v-list-item-title
-						class="text-wrap"
-						v-text="homework.name"
-					></v-list-item-title>
-				</v-list-item-content>
-				<v-list-item-action>
-					<v-list-item-action-text
-						class="subtitle-2"
-						data-test-id="dueDateLabel"
-						v-text="
-							computedDueDateLabel(
-								homework.duedate,
-								(shorten = $vuetify.breakpoint.xsOnly)
-							)
-						"
-					/>
-					<v-spacer />
-					<v-custom-chip-time-remaining
-						v-if="isCloseToDueDate(homework.duedate)"
-						type="warning"
-						:due-date="homework.duedate"
-						:shorten-date="$vuetify.breakpoint.xsOnly"
-					/>
-					<v-custom-chip-time-remaining
-						v-else-if="isOverDue(homework.duedate)"
-						type="overdue"
-						:due-date="homework.duedate"
-						:shorten-date="$vuetify.breakpoint.xsOnly"
-					/>
-				</v-list-item-action>
-			</v-list-item>
-			<v-divider v-if="index < homeworks.length - 1" :key="index"></v-divider>
+			<v-homework-item-student
+				v-if="type === 'student'"
+				:key="index"
+				:homework="homework"
+			/>
+			<v-divider
+				v-if="index < homeworks.length - 1"
+				:key="`divider-${index}`"
+			/>
 		</template>
 	</v-list>
 </template>
 
 <script>
-import VCustomChipTimeRemaining from "@components/molecules/VCustomChipTimeRemaining";
-import { fromNow, fromNowToFuture } from "@plugins/datetime";
-import taskIconSvg from "@assets/img/courses/task-new.svg";
-import {
-	printDateFromStringUTC,
-	printDateTimeFromStringUTC,
-} from "@plugins/datetime";
+import VHomeworkItemStudent from "@components/molecules/vHomeworkItemStudent";
 import { mapGetters } from "vuex";
 
 export default {
-	components: { VCustomChipTimeRemaining },
+	components: { VHomeworkItemStudent },
 	props: {
 		homeworks: {
 			type: Array,
@@ -81,12 +44,11 @@ export default {
 			required: false,
 			default: null,
 		},
-	},
-	data() {
-		return {
-			fromNow,
-			taskIconSvg,
-		};
+		type: {
+			type: String,
+			required: true,
+			validator: (value) => ["student", "teacher"].includes(value),
+		},
 	},
 	computed: {
 		...mapGetters("homeworks", {
@@ -94,35 +56,8 @@ export default {
 		}),
 	},
 	methods: {
-		computedDueDateLabel(dueDate, shorten = false) {
-			if (!dueDate) {
-				return this.$t("pages.homeworks.labels.noDueDate");
-			} else if (shorten) {
-				return (
-					this.$t("pages.homeworks.labels.due") +
-					printDateFromStringUTC(dueDate)
-				);
-			} else {
-				return (
-					this.$t("pages.homeworks.labels.due") +
-					printDateTimeFromStringUTC(dueDate)
-				);
-			}
-		},
-		isCloseToDueDate(dueDate) {
-			const timeDiff = fromNowToFuture(dueDate, "hours");
-			if (timeDiff === null) {
-				return false;
-			} else return timeDiff <= 24;
-		},
 		isListFilled() {
 			return this.loading === false && this.homeworks.length > 0;
-		},
-		isOverDue(dueDate) {
-			return dueDate && new Date(dueDate) < new Date();
-		},
-		homeworkHref: (id) => {
-			return "/homework/" + id;
 		},
 	},
 };
